[SW-1937] Move RestAPIUtils to ai.h2o.sparkling package
