+++ core/src/main/scala/ai/h2o/sparkling/utils/H2OContextUtils.scala
@@ -1,0 +1,48 @@
+/*
+* Licensed to the Apache Software Foundation (ASF) under one or more
+* contributor license agreements.  See the NOTICE file distributed with
+* this work for additional information regarding copyright ownership.
+* The ASF licenses this file to You under the Apache License, Version 2.0
+* (the "License"); you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+package ai.h2o.sparkling.utils
+
+import java.io.File
+import java.text.SimpleDateFormat
+import java.util.Date
+
+import org.apache.spark.h2o.H2OConf
+import org.apache.spark.h2o.backends.external.RestCommunication
+
+trait H2OContextUtils extends RestCommunication with RestApiUtils {
+
+  private def logFileName(): String = {
+    val pattern = "yyyyMMdd_hhmmss"
+    val formatter = new SimpleDateFormat(pattern)
+    val now = formatter.format(new Date)
+    s"h2ologs_$now"
+  }
+
+  def downloadLogs(destinationDir: String, logContainer: String, conf: H2OConf): String = {
+    val endpoint = RestApiUtils.getClusterEndpoint(conf)
+    val file = new File(destinationDir, s"${logFileName()}.${logContainer.toLowerCase}")
+    val logEndpoint = s"/3/Logs/download/$logContainer"
+    logContainer match {
+      case "LOG" =>
+        downloadStringURLContent(endpoint, logEndpoint, conf, file)
+      case "ZIP" =>
+        downloadBinaryURLContent(endpoint, logEndpoint, conf, file)
+    }
+    file.getAbsolutePath
+  }
+}
