+++ core/src/main/scala/org/apache/spark/h2o/backends/external/ProxyStarter.scala
@@ -19,6 +19,7 @@ package org.apache.spark.h2o.backends.external
 
 import java.net._
 
+import ai.h2o.sparkling.utils.RestApiUtils
 import org.apache.spark.SparkEnv
 import org.apache.spark.expose.Logging
 import org.apache.spark.h2o.H2OConf
