+++ ml/src/main/scala/ai/h2o/sparkling/ml/algos/H2OKMeans.scala
@@ -18,10 +18,10 @@ package ai.h2o.sparkling.ml.algos
 
 import ai.h2o.sparkling.frame.{H2OColumnType, H2OFrame}
 import ai.h2o.sparkling.ml.params.{H2OAlgoParamsHelper, H2OAlgoUnsupervisedParams}
+import ai.h2o.sparkling.utils.RestApiUtils
 import hex.kmeans.KMeansModel.KMeansParameters
 import hex.kmeans.{KMeans, KMeansModel}
 import hex.schemas.GLMV3.GLMParametersV3
-import org.apache.spark.h2o.backends.external.RestApiUtils
 import org.apache.spark.h2o.{Frame, H2OContext}
 import org.apache.spark.ml.util.{DefaultParamsReadable, Identifiable}
 import org.apache.spark.sql.SparkSession
